// types.ts
export type InventoryProduct = {
  id: string;
  name: string;
  quantity: number;
  sellingPrice: number;
};

export type SellType = {
  id?:string;
  inventoryProduct: string;
  quantity: number;
  Price: number;
  loan?: boolean;
  loanAmount?: number;
  PriceAfterLoan: number;
  discount?: number;
  date: Date | string | Timestamp;
  customerName?: string;
  customerNumber?: number;
  description?: string;
  ownerName?: string;
};